Share code to compile libtest and require it from benchmarks too
